--- pymodbus/transport/serial_asyncio/README.md
@@ -1,3 +1,0 @@
-Patch for serial_asyncio, also submitted to the repo:
-
-https://github.com/pyserial/pyserial-asyncio/pull/94
